bug 1277577 - Fix Template edit test
The test test_show_toc_hidden_input_for_template didn't actually test
what it claimed to test - that the toc_depth input element was hidden
and had a value of 0.  Instead, it was testing the hidden inputs on the
Template display page, without checking that the desired element was
actually there.

Moved the test to DocumentEditTests, and switched to verifying the
toc_depth element using PyQuery. This also removes the last use of
BeautifulSoup in the codebase, so removed the dependency.